# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для управления студентами, классами и лекциями",
        "version": "1.0"
    },
    "paths": {
...
        "/api/lectures": {
            "get": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Получить список всех лекций",
                "operationId": "91863ee34b872ea123355cc2423a40da",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Lecture"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Создать лекцию",
                "operationId": "9b39872f2fe89a73d8bb84091e9ddf0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLectureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Лекция создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lecture"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/lectures/{id}": {
            "get": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Получить информацию о конкретной лекции",
                "operationId": "0b3e6ad31febe5d6c9b241b3f35ff39e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "topic": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "classes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "order": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "students": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "class": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "listened_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Лекция не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Обновить лекцию",
                "operationId": "e9518d236b3e11e73f306c98a0f90dc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLectureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Лекция обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lecture"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Лекция не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lectures"
                ],
                "summary": "Удалить лекцию",
                "operationId": "9428f55db7e50cc1ae3cd2ab92adc77b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Лекция удалена"
                    },
                    "404": {
                        "description": "Лекция не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreLectureRequest": {
                "required": [
                    "topic",
                    "description"
                ],
                "properties": {
                    "topic": {
                        "type": "string",
                        "example": "Математика: Алгебра"
                    },
                    "description": {
                        "type": "string",
                        "example": "Основы алгебраических выражений"
                    }
                },
                "type": "object"
            },
            "UpdateLectureRequest": {
                "properties": {
                    "topic": {
                        "type": "string",
                        "example": "Математика: Алгебра"
                    },
                    "description": {
                        "type": "string",
                        "example": "Основы алгебраических выражений"
                    }
                },
                "type": "object"
            },
...
            "Lecture": {
                "title": "Lecture",
                "description": "Модель лекции",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "topic": {
                        "type": "string",
                        "example": "Математика: Алгебра"
                    },
                    "description": {
                        "type": "string",
                        "example": "Основы алгебраических выражений"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}